refactor: 아코디언 컴포넌트 분리, 훅 분리
